fix: escape special characters in test name
Keep the double quotes in the test name argument as some test runners are using Regex to find the test to run. If we use the single quotes to
avoid escaping the special characters, those test runners will interpre
t the special characters as operators in their Regex.

We are escaping all special characters and the double quotes to make su
re the CodeLens can generate a working command in the cli.